Move the helper for computing StatisticResults inside the type
